Added Student Website to SmartNoiseSense (#20)
* Made Website Shell

Easy website has been made, next is to add firebase functions

* Added Linter File

This helps enforce consistent styling accross the code for the website.
Although a small project, style matters!

* Added simple website components

* Added Firebase Link to Webstie

Now we are able to grab data published by the sound sensors. All that's
left is to be able to visualize the data.

* Added Helper Function to Get Day and Hour

Now we have some functions to call to easily convert a day to a string,
as well as the hour to a string. This comes in handy when trying to
query firestore. I'm on a train right now

* Added Listener for New Data Added

* Added component to display data from listener

* Added Default values For Parameters

I added a default value for the parameters to be the current day and
hour for getting sound data. These parameters can be something else when
we want to look at analytics.

* Added Component to Show Sound by Minute

This component will return the sound for every minute, and if no reading
exists from the hour, the component will tell the user to wait at least
for one minute for a reading.

* Added Beginning of Module to Show Analytics

I added a dropdown to specify the day, the next moves are to send this
change to our firebaseActions to get the day and hour for each time.

* Moved Components to Their Own Files

This promotes readability for those looking at our projects

* Moved DropDown Menu to its Own File

Promotes scalability.

* Added Title and Logo

* Changed Spacing of Message

Now, the messages will be NE:<sound> with no space after colon. Helps
parse with cloud function.

* Changed Reading Firebase to Get Hourly Average

Now we are getting a value for each of the hours instead of all of the
hourly readings from firebase. This helps as we just show what the
overall average was, as well as the average for each room by hour.

* Added Graph to StudentWebsite

Now we can see the average plotted overtime. Yay

* changed average to have 2 decimals

* Added Basic Heatmap

All that is modified is the text color. I will soon look into how to
render better colors over time.